+++ server/resources/user/user.router.ts
@@ -4,7 +4,7 @@ import {
   addUser,
   updateUser,
   deleteUser,
-  handleLogin,
+  // handleLogin,
   // handleLogout,
 } from './user.controller';
 
@@ -14,5 +14,5 @@ export const userRouter = express
   .post('/user/register', addUser)
   .put('/user/:id', updateUser)
   .delete('/user/:id', deleteUser)
-  .post('/user/login', handleLogin)
+  // .post('/user/login', handleLogin)
   // .post('/user/logout', handleLogout)
